Fix unfitting types in function signatures
All callers of `updateGain` already have an `int`, except for
`setChannelVolume` where the single caller also has an `int`.

The only caller of `getInternalOutputChannelName` has `size_t` and the
function itself works with any numeric type.

The `entryIndex` is always an `int` in the caller and the called
functions.

Changed these function signatures to prevent warnings about integer
conversions with possible loss of data.

+++ Source/OutputChannelNames.h
@@ -65,7 +65,7 @@ private:
 public:
     juce::String getDeviceOutputChannelName(int activeChannelIndex);
     juce::StringArray getAllDeviceOutputChannelNames();
-    juce::String getInternalOutputChannelName(int activeChannelIndex);
+    juce::String getInternalOutputChannelName(size_t activeChannelIndex);
     void setInternalOutputChannelName(int activeChannelIndex, const juce::String& text);
 
 private:
